add realtime heatmap plotting scripts

+++ realtime/simulation.py
@@ -2,4 +2,4 @@ import numpy as np
 
 
 def sim_heatmap(shape: list):
-    return np.random.random(shape)
+    return np.random.random(shape).astype(np.float16)
